requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
geoalchemy2==0.15.2
psycopg2==2.9.10
python-dotenv==1.0.1
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
shapely==2.0.6
openpyxl==3.1.5
